+++ src/drivers/i2c.c
@@ -33,7 +33,7 @@
 #if CONFIG_I2C == 1
 
 struct module_t {
-    int initialized;
+    int8_t initialized;
 #if CONFIG_I2C_FS_COMMAND_READ == 1
     struct fs_command_t cmd_read;
 #endif
